change column for flight record

+++ src/main/java/com/pigeon_management_system_api/dto/FlightRecordDTO.java
@@ -14,7 +14,7 @@ public class FlightRecordDTO {
     @NotEmpty
     private String pigeonRing;
     @NotEmpty
-    private Double position;
+    private Integer position;
     @NotEmpty
     private Double coefic;
     @NotEmpty
