+++ src/main/java/com/pigeon_management_system_api/model/FlightRecord.java
@@ -18,7 +18,7 @@ public class FlightRecord implements Serializable {
     private Integer flightId;
     @Column(name = "pigeon_ring")
     private String pigeonRing;
-    private Double position;
+    private Integer position;
     private Double coefic;
     private Double points;
 }
